fix some small bugs on the c++ interface

+++ src/interface_library/comm_agent.h
@@ -13,8 +13,8 @@ class CommAgent {
 public:
     CommAgent(const std::string &serverUrl, std::shared_ptr<zmq::context_t> ctxPtr=nullptr);
     ~CommAgent();
-    float getUtilityThreshold(float dropRatio);
-    float getUtilityValue(Features::Builder &utilityRequest);
+    double getUtilityThreshold(float dropRatio);
+    double getUtilityValue(Features::Builder &utilityRequest);
 private:
     std::shared_ptr<zmq::context_t> ctx;
     std::unique_ptr<zmq::socket_t> sock;
